Move RulesProviderSpy class to its own file
